+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- Very simple parser, which should provide just enough of a basis to add basic editor functionality like code
+  formatting and auto indentation.
+
 ## [0.1.3-alpha] - 2023-03-30
 
 ### Changed